refactor: migrate units-editor to bundled TypeScript controller
Port public/modules/ui/units-editor.js to src/controllers/units-editor.ts
following the migration guide (strict d3 v7: select/drag, event-arg ruler
gestures). The #unitsEditor markup stays in index.html because its inputs
(distanceUnitInput, heightUnit, temperatureScale, …) are app-wide settings
cached as globals at load; listeners are wired once behind an initialized
flag. Register as Controllers.UnitsEditor and replace editUnits() at the
main.js, load.ts, tools.js and hotkeys.js call sites; drop the editUnits
global. Add toggleRulers, calculateFriendlyGridSize, RouteOpisometer globals.
