+++ public/modules/ui/tools.js
@@ -20,7 +20,7 @@ toolsContent.addEventListener("click", function (event) {
   else if (button === "editGoods") window.Controllers.GoodsEditor.open();
   else if (button === "editEmblemButton") openEmblemEditor();
   else if (button === "editNamesBaseButton") window.Controllers.NamesbaseEditor.open();
-  else if (button === "editUnitsButton") editUnits();
+  else if (button === "editUnitsButton") window.Controllers.UnitsEditor.open();
   else if (button === "editNotesButton") window.Controllers.NotesEditor.open();
   else if (button === "editZonesButton") editZones();
   else if (button === "overviewChartsButton") window.Controllers.ChartsOverview.open();
