+++ src/index.html
@@ -8577,7 +8577,6 @@
     <script defer src="modules/ui/coastline-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/labels-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/relief-editor.js?v=1.99.00"></script>
-    <script defer src="modules/ui/units-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/ai-generator.js?v=1.120.5"></script>
     <script defer src="modules/ui/diplomacy-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/zones-editor.js?v=1.120.5"></script>
